Descripción de los cambios

+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -7,7 +7,7 @@ import {CartsService} from '../../services/cart.service';
 import {MatDivider} from "@angular/material/divider";
 import {MatButton, MatIconButton} from "@angular/material/button";
 import {Router, RouterLink} from "@angular/router";
-import {ReactiveFormsModule} from "@angular/forms";
+import {FormsModule, ReactiveFormsModule} from "@angular/forms";
 import {MatIcon} from "@angular/material/icon";
 
 @Component({
@@ -15,7 +15,7 @@ import {MatIcon} from "@angular/material/icon";
   templateUrl: './cart-list.component.html',
   styleUrls: ['./cart-list.component.css'],
   standalone: true,
-  imports: [MatCard, MatCardModule, NgForOf, MatDivider, MatButton, RouterLink, ReactiveFormsModule, MatIcon, MatIconButton],
+  imports: [MatCard, MatCardModule, NgForOf, MatDivider, MatButton, RouterLink, ReactiveFormsModule, MatIcon, MatIconButton, FormsModule],
 })
 export class CartListComponent {
   cartItems: Cart[] = [];
@@ -31,19 +31,41 @@ export class CartListComponent {
       this.cartItems = data;
     });
   }
-  removeFromCart(itemId: number): void {
-    this.cartService.removeFromCart(itemId).subscribe(() => {
-      // Actualizar la lista de elementos del carrito después de eliminar un elemento
-      this.getCartItems();
-    });
-  }
-  calculateTotal() {
-    // Calcular el total de la compra sumando los precios de los productos en el carrito
-    let total = 0;
-    for (const item of this.cartItems) {
-      total += item.price;
+  // cart-list.component.ts
+updateQuantity(itemId: number, quantity: number): void {
+  this.cartService.updateQuantity(itemId, quantity).subscribe((updatedItem: Cart) => {
+    // Actualiza la cantidad del producto en el carrito
+    const item = this.cartItems.find(item => item.id === itemId);
+    if (item) {
+      item.quantity = updatedItem.quantity;
     }
-    return total;
-  }
+  });
+}
 
+calculateTotal() {
+  let total = 0;
+  for (const item of this.cartItems) {
+    // Asegúrate de que la cantidad siempre sea un número
+    const quantity = Number(item.quantity) || 0;
+    total += item.price * quantity;
+  }
+  return total;
+}
+
+removeFromCart(itemId: number): void {
+  const item = this.cartItems.find(item => item.id === itemId);
+  if (item) {
+    if (item.quantity > 1) {
+      // Si la cantidad del producto es mayor que 1, disminuye la cantidad
+      this.updateQuantity(itemId, item.quantity - 1);
+    } else {
+      // Si la cantidad del producto es 1, elimina el producto del carrito
+      this.cartService.removeFromCart(itemId).subscribe(() => {
+        this.getCartItems();
+      });
+    }
+  }
+}
+  protected readonly HTMLInputElement = HTMLInputElement;
+  protected readonly Number = Number;
 }
